# Render marshalSJON/Raw/CSV preallocated bytes buffer