```
$ go test -benchmem -run=^$ -bench ^Benchmark github.com/go-graphite/carbonapi/expr/types | tee marshalJSON2.txt 

$ benchcmp marshalJSON1.txt marshalJSON2.txt 
benchmark                      old ns/op     new ns/op     delta
BenchmarkMarshalJSON-6         9297315       7406833       -20.33%
BenchmarkMarshalJSONLong-6     135895934     121057103     -10.92%
BenchmarkMarshalRaw-6          5717379       5622104       -1.67%
BenchmarkMarshalRawLong-6      94664746      90235233      -4.68%
BenchmarkMarshalCSV-6          38516100      19187483      -50.18%
BenchmarkMarshalCSVLong-6      587776864     320056366     -45.55%

benchmark                      old allocs     new allocs     delta
BenchmarkMarshalJSON-6         39             5              -87.18%
BenchmarkMarshalJSONLong-6     89             43             -51.69%
BenchmarkMarshalRaw-6          29             1              -96.55%
BenchmarkMarshalRawLong-6      42             1              -97.62%
BenchmarkMarshalCSV-6          110037         3              -100.00%
BenchmarkMarshalCSVLong-6      1820049        6              -100.00%

benchmark                      old bytes     new bytes     delta
BenchmarkMarshalJSON-6         8401270       2569729       -69.41%
BenchmarkMarshalJSONLong-6     130283752     27232828      -79.10%
BenchmarkMarshalRaw-6          1981197       2564105       +29.42%
BenchmarkMarshalRawLong-6      41704200      25608206      -38.60%
BenchmarkMarshalCSV-6          23749513      9773060       -58.85%
BenchmarkMarshalCSVLong-6      358308832     288342088     -19.53%

```