@msaf1980 2022-10-20
Also may be better direct write to http.ResponseWriter instead of use bytes buffer. It's will be require less memory for large metrrics set.