@Civil 2022-10-20
Yeah, that should work as well. I'm not sure it won't have any complications on the behavior, but might improve time to first byte as well for consumers.