@msaf1980 2022-10-20
> Yeah, that should work as well. I'm not sure it won't have any complications on the behavior, but might improve time to first byte as well for consumers.

Also  I'm not have experience with this. It can produce chunked response (on large metrics set) and need more tests, that no bugs in http clients, grafana, etc.
May be better merge current, and research some later ?